pg_dump/requirements.txt
python-dotenv==1.0.0
pydantic==1.10.9
psycopg2-binary==2.9.9